Selection_Sort_Visualisation.cpp
The human brain can easily process visuals in spite of long codes to understand the algorithms. In this article, Selection Sort visualization has been implemented using graphics.h library. As we all know selection sort first finds the minimum element from the unsorted array and swaps it with the first element of the unsorted array in each pass. It becomes difficult to analyze the data manually between two algorithms or vice versa, but plotting graphically it is much easier to understand.